requirements.txt
Flask==3.0.0
flask-cors==4.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
PyYAML==6.0.1

